Ropt: move init implementation
