+++ src/main/java/org/apache/sysds/api/ropt/old_impl/CloudInstanceAnalyzer.java
@@ -1,4 +1,4 @@
-package org.apache.sysds.api.ropt;
+package org.apache.sysds.api.ropt.old_impl;
 
 import scala.xml.dtd.ValidationException;
