+++ src/main/java/org/apache/sysds/api/ropt/old_impl/CloudInstanceConfig.java
@@ -1,4 +1,4 @@
-package org.apache.sysds.api.ropt;
+package org.apache.sysds.api.ropt.old_impl;
 
 /**
  * This class describes the configurations of a single VM instance.
