+++ src/main/java/org/apache/sysds/api/ropt/old_impl/EMRUtils.java
@@ -1,4 +1,4 @@
-package org.apache.sysds.api.ropt;
+package org.apache.sysds.api.ropt.old_impl;
 
 import java.util.EnumSet;
 import java.util.Set;
